+++ src/resolver/index.ts
@@ -1,3 +1,4 @@
 export { extractDependencies, extractAllDependencies } from './extract-dependencies';
-export { applyWrappers as wrap } from './apply-wrappers';
+export { applyWrappers } from './apply-wrappers';
 export { createWrapper } from './create-wrapper';
+export { WrapperFor } from './constants';
