--- src/utils/index.ts
@@ -1,1 +1,0 @@
-export { walk } from './walk';
